Bug fixes and tweaks
Room IDs are now properly chosen (oops)

Files are now laid out in an acceptable way. I might improve it at some point to make it a centred grid, and also make it scale properly for mobile.

Files now hide the remaining time while downloading. The time is also reset to the full length when a download fails

The correct pointer is now used for buttons and sans-serif is now used.